Move AIO test to it's own file
It was causing a crash because of conflict with sync tests
