+++ ftd2xx/tests/t_ftd2xx.py
@@ -226,28 +226,5 @@ class TestGlobalFunctions(unittest.TestCase):
             dev0.close()
 
 
-class TestAIOFTD2XX(TestFTD2XX, unittest.IsolatedAsyncioTestCase):
-    def setUp(self):
-        self.device = aio.open()
-
-    async def testread(self):
-        self.device.setTimeouts(1000, 0)
-        result = await self.device.read(1)
-        self.assertTrue(isinstance(result, bytes))
-
-    async def testreadtimeout(self):
-        self.device.setTimeouts(1, 0)
-
-        for _ in range(100):
-            try:
-                await self.device.read(1, exc=True)
-                self.device.purge()
-                continue
-            except Exception as e:
-                self.assertTrue(isinstance(e, asyncio.TimeoutError))
-                return
-        raise Exception("Could not timeout!")
-
-
 if __name__ == "__main__":
     unittest.main()
